FICHEROS/requirements.txt
google-generativeai
python-dotenv
pillow
openpyxl
